Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/iframe-blocked-when-loaded-via-javascript-url2.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self' 'unsafe-inline'">
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<script>
var test = "javascript:'<iframe src=\"javascript:`<iframe src=http://localhost:8000/security/contentSecurityPolicy/resources/alert-fail.html></iframe><script>window.testRunner && window.setTimeout(() => testRunner.notifyDone(), 0);</" + "script>`\"></iframe>'";
window.open(test, "_self");
</script>
</body>
</html>
